% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lwia_data.R
\name{combine_lwia_data}
\alias{combine_lwia_data}
\title{Combine lgr data}
\usage{
combine_lwia_data(parsed_data, file_rownames = FALSE)
}
\arguments{
\item{parsed_data}{an output from parse_lwia_files or trim_lwia_data}

\item{file_rownames}{logical TRUE/FALSE to include file names as ronwames}
}
\description{
Combine several parsed files into a single data frame for subsequent analysis.
}
