% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standards_library.R
\name{get_standards_library}
\alias{get_standards_library}
\title{Get a standards library}
\usage{
get_standards_library(directory, name = "LGRLibrary")
}
\arguments{
\item{directory}{a folder to search inside}

\item{name}{A character string to use for the hunt.}
}
\description{
Find a LGR analyser standards library file
}
