% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lwia_files.R
\name{parse_lwia_files}
\alias{parse_lwia_files}
\title{Load LGR LWIA post-processing files into the workspace.}
\usage{
parse_lwia_files(target, type = "Processed", rc = 0)
}
\arguments{
\item{target}{A complete folder or file path as a string.}

\item{type}{a group of EEMs compliant with the staRdom/EEM/eemR package framework}

\item{rc}{used by the function to tell if it is recursing (0 for no, 1 for yes).}
}
\description{
Read and parse the two file types generated when saving processed data from the LWIA post-processing software
}
