% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_lwia_data.R
\name{trim_lwia_data}
\alias{trim_lwia_data}
\title{Trim one or more parsed LWIA data files}
\usage{
trim_lwia_data(
  parsed_data,
  remove_DI = c("DI", "DIS"),
  remove_externals = c("VSMOW2", "VMSOW", "GRESP", "SLAP", "SLAP2"),
  remove_internals = NULL
)
}
\arguments{
\item{parsed_data}{output from parse_lwia_files.}

\item{remove_DI}{NULL or character vector with DI water sample/control/standard strings.}

\item{remove_externals}{NULL or character vector with external calibration sample/control/standard strings.}

\item{remove_internals}{NULL or character vector with internal calibration sample/control/standard strings.}
}
\description{
Trim one or more LWIA data files parsed into R by removing unwanted measurement types. This performs a lapply
wrap around the internal function clean_data.
}
